requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.1
frozenlist==1.2.0
fsspec==2022.1.0
gcsfs==2022.1.0
google-api-core==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
multidict==5.2.0
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
urllib3==1.26.8
yarl==1.7.2
openpyxl
pandas-gbq
emoji
dropbox==11.31.0
bs4
requests_html
html5lib
XlsxWriter